deployment/web_service_local/requirements.txt
tf-nightly
mlflow
numpy
scikit-learn
gunicorn
google-cloud-storage
python-dotenv
boto3
flask
gcloud
cryptography
dill
fsspec
psycopg2-binary
google-api-python-client
